review comments

+++ ewm-service/src/main/java/ru/practicum/ewm/dto/event/EventShortDto.java
@@ -16,11 +16,11 @@ public class EventShortDto {
     private Long id;
     private String annotation;
     private CategoryDto category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
     private UserShortDto initiator;
     private Boolean paid;
     private String title;
-    private Integer views;
+    private Long views;
 }
